requirements.txt
ipykernel
langchain
langgraph
langchain-core
langchain-community
langchain_core
langchain_community
python-dotenv
langchain-groq
langsmith
langchain-text-splitters
langchain_openai
pydantic
arxiv
wikipedia
tavily-python
faiss-cpu
langchain_ollama
langchain_huggingface
sentence_transformers
streamlit
pypdf
langchain_chroma
ipython
langchain-tavily